EWPP-4314: Check if locale enabled in update path.

+++ oe_corporate_blocks.post_update.php
@@ -336,11 +336,13 @@ function oe_corporate_blocks_post_update_40010(&$sandbox): void {
   // Allow for config translation re-import when running
   // "drush oe-multilingual:import-local-translations".
   // @see https://citnet.tech.ec.europa.eu/CITnet/jira/browse/OPENEUROPA-2407
-  $storage = \Drupal::service('locale.storage');
-  $string = $storage->findString(['source' => 'Search all EU institutions and bodies']);
-  $storage->delete($string);
-  $string = $storage->findString(['source' => 'EU institutions and bodies']);
-  $storage->delete($string);
-  $string = $storage->findString(['source' => 'https://european-union.europa.eu/institutions-law-budget/institutions-and-bodies/search-all-eu-institutions-and-bodies_en']);
-  $storage->delete($string);
+  if (\Drupal::moduleHandler()->moduleExists('locale')) {
+    $storage = \Drupal::service('locale.storage');
+    $string = $storage->findString(['source' => 'Search all EU institutions and bodies']);
+    $storage->delete($string);
+    $string = $storage->findString(['source' => 'EU institutions and bodies']);
+    $storage->delete($string);
+    $string = $storage->findString(['source' => 'https://european-union.europa.eu/institutions-law-budget/institutions-and-bodies/search-all-eu-institutions-and-bodies_en']);
+    $storage->delete($string);
+  }
 }
